for-deletion machines dont want for replica change
